chore: consolidate all runtime sheet into the same file to enable more agressive filesize reductions
