+++ packages/core/src/__tests__/browser.test.tsx
@@ -93,13 +93,13 @@ describe('browser', () => {
     render(<StyledLink href="https://atlassian.design">Atlassian Design System</StyledLink>);
 
     expect(document.head.innerHTML.split('</style>').join('</style>\n')).toMatchInlineSnapshot(`
-      "<style nonce=\\"k0Mp1lEd\\">._1e0c1txw{display:flex}._1wyb12am{font-size:50px}._syaz1cnh{color:purple}</style>
-      <style nonce=\\"k0Mp1lEd\\" data-c=\\"l\\">._ysv75scu:link{color:red}._ysv71x77:link, ._v0vw1x77:focus-visible{color:white}</style>
-      <style nonce=\\"k0Mp1lEd\\" data-c=\\"v\\">._105332ev:visited{color:pink}</style>
-      <style nonce=\\"k0Mp1lEd\\" data-c=\\"f\\">._f8pjbf54:focus{color:green}</style>
-      <style nonce=\\"k0Mp1lEd\\" data-c=\\"h\\">._30l31gy6:hover{color:yellow}</style>
-      <style nonce=\\"k0Mp1lEd\\" data-c=\\"a\\">._9h8h13q2:active{color:blue}</style>
-      <style nonce=\\"k0Mp1lEd\\" data-c=\\"m\\">@supports (display: grid){._qxnw1gy6:focus{color:yellow}._p7tw11x8:active{color:black}}@media (max-width: 800px){._vyxz1gy6:focus{color:yellow}._7zyjtwqo:hover, ._ov6atwqo:focus-visible{color:grey}._ojvu11x8:active{color:black}}</style>
+      "<style nonce=\\"k0Mp1lEd\\">._1e0c1txw{display:flex}._1wyb12am{font-size:50px}._syaz1cnh{color:purple}._ysv71x77:link, ._v0vw1x77:focus-visible{color:white}</style>
+      <style nonce=\\"k0Mp1lEd\\">._ysv75scu:link{color:red}</style>
+      <style nonce=\\"k0Mp1lEd\\">._105332ev:visited{color:pink}</style>
+      <style nonce=\\"k0Mp1lEd\\">._f8pjbf54:focus{color:green}</style>
+      <style nonce=\\"k0Mp1lEd\\">._30l31gy6:hover{color:yellow}</style>
+      <style nonce=\\"k0Mp1lEd\\">._9h8h13q2:active{color:blue}</style>
+      <style nonce=\\"k0Mp1lEd\\">@supports (display: grid){._qxnw1gy6:focus{color:yellow}._p7tw11x8:active{color:black}}@media (max-width: 800px){._vyxz1gy6:focus{color:yellow}._7zyjtwqo:hover, ._ov6atwqo:focus-visible{color:grey}._ojvu11x8:active{color:black}}</style>
       "
     `);
   });
