+++ packages/core/src/__tests__/ssr.test.tsx
@@ -122,13 +122,7 @@ describe('SSR', () => {
     );
 
     expect(result.split('</style>').join('</style>\n')).toMatchInlineSnapshot(`
-      "<style nonce=\\"k0Mp1lEd\\">._1e0c1txw{display:flex}._1wyb12am{font-size:50px}._syaz1cnh{color:purple}</style>
-      <style data-c=\\"l\\" nonce=\\"k0Mp1lEd\\">._ysv75scu:link{color:red}</style>
-      <style data-c=\\"v\\" nonce=\\"k0Mp1lEd\\">._105332ev:visited{color:pink}</style>
-      <style data-c=\\"f\\" nonce=\\"k0Mp1lEd\\">._f8pjbf54:focus{color:green}</style>
-      <style data-c=\\"h\\" nonce=\\"k0Mp1lEd\\">._30l31gy6:hover{color:yellow}</style>
-      <style data-c=\\"a\\" nonce=\\"k0Mp1lEd\\">._9h8h13q2:active{color:blue}</style>
-      <style data-c=\\"m\\" nonce=\\"k0Mp1lEd\\">@supports (display: grid){._qxnw1gy6:focus{color:yellow}._p7tw11x8:active{color:black}}@media (max-width: 800px){._vyxz1gy6:focus{color:yellow}._ojvu11x8:active{color:black}}</style>
+      "<style nonce=\\"k0Mp1lEd\\">._1e0c1txw{display:flex}._1wyb12am{font-size:50px}._syaz1cnh{color:purple}._ysv75scu:link{color:red}._105332ev:visited{color:pink}._f8pjbf54:focus{color:green}._30l31gy6:hover{color:yellow}._9h8h13q2:active{color:blue}@supports (display: grid){._qxnw1gy6:focus{color:yellow}._p7tw11x8:active{color:black}}@media (max-width: 800px){._vyxz1gy6:focus{color:yellow}._ojvu11x8:active{color:black}}</style>
       <a href=\\"https://atlassian.design\\" class=\\"_1e0c1txw _1wyb12am _syaz1cnh _30l31gy6 _9h8h13q2 _ysv75scu _p7tw11x8 _qxnw1gy6 _f8pjbf54 _105332ev _ojvu11x8 _vyxz1gy6\\">Atlassian Design System</a>"
     `);
   });
